:construction: Move login code from scrape.py to login.py
